requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
drf-yasg==1.21.7
Faker==24.3.0
inflection==0.5.1
Markdown==3.6
packaging==24.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
validate-docbr==1.10.0
